+++ jni/lib/src/util/jmap.dart
@@ -107,11 +107,9 @@ class JMap<$K extends JObject, $V extends JObject> extends JObject
     if (other is JMap<$K, $V>) {
       Jni.accessors.callMethodWithArgs(reference, _addAllId,
           JniCallType.voidType, [other.reference]).check();
-    } else {
-      for (final entry in other.entries) {
-        this[entry.key] = entry.value;
-      }
+      return;
     }
+    super.addAll(other);
   }
 
   static final _clearId =